use MetaStoreDirectSql

+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/HiveAlterHandler.java
@@ -202,7 +202,7 @@ public class HiveAlterHandler implements AlterHandler {
           throw new MetaException("The table has been modified. The parameter value for key '" + expectedKey + "' is '"
               + oldt.getParameters().get(expectedKey) + "'. The expected was value was '" + expectedValue + "'");
         }
-        int affectedRows = msdb.updateParameterWithExpectedValue(oldt, expectedKey, expectedValue, newValue);
+        long affectedRows = msdb.updateParameterWithExpectedValue(oldt, expectedKey, expectedValue, newValue);
         if (affectedRows != 1) {
           throw new MetaException(String.format(
               "The table has been modified. Updating expected key %s affects %d rows", expectedKey, affectedRows));
